backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.9
langchain-openai==0.0.8
langgraph==0.0.26
langchain-community==0.0.24
openai==1.12.0
pydantic==2.6.1
pydantic-settings==2.2.1
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
chromadb==0.4.22
pandas==2.2.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiohttp==3.9.3
pytest==8.0.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1
boto3==1.34.34
tiktoken==0.6.0
faiss-cpu==1.7.4
yfinance==0.2.36
newsapi-python==0.2.7
slowapi==0.1.9
